added core structure to the app
