+++ apps/frontend/src/App.css
@@ -1,0 +1,7 @@
+.card-list {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 1rem;
+}
