--- apps/frontend/src/App.tsx
@@ -1,16 +1,0 @@
-import * as React from "react";
-import "./App.css";
-import { Button } from "ui";
-
-function App() {
-  return (
-    <div className="container">
-      <h1 className="title">
-        Helping Hand
-        <Button />
-      </h1>
-    </div>
-  );
-}
-
-export default App;
